@@ -15,53 +15,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.5",
+    "@angular/animations": "^8.2.9",
     "@angular/cli": "^8.3.7",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/common": "^8.2.9",
+    "@angular/compiler": "^8.2.9",
+    "@angular/compiler-cli": "^8.2.9",
+    "@angular/core": "^8.2.9",
+    "@angular/forms": "^8.2.9",
+    "@angular/material": "^8.2.2",
+    "@angular/platform-browser": "^8.2.9",
+    "@angular/platform-browser-dynamic": "^8.2.9",
+    "@angular/router": "^8.2.9",
+    "@auth0/angular-jwt": "^3.0.0",
     "@ng-bootstrap/ng-bootstrap": "^1.0.1",
-    "angular-4-data-table": "^0.4.6",
-    "angular2-jwt": "^0.2.3",
-    "angular2-uuid": "^1.1.1",
-    "angular5-data-table": "^0.5.1",
-    "bootstrap": "^4.0.0",
+    "bootstrap": "^4.3.1",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "jquery": "^3.3.1",
+    "jquery": "^3.4.1",
     "ng2-slider-component": "^1.0.9",
     "ngx-bootstrap": "^2.0.2",
     "ngx-gallery": "^5.3.2",
     "ngx-toastr": "^8.2.1",
     "popper": "^1.0.1",
-    "primeng": "^5.2.1",
-    "rxjs": "^5.5.6",
+    "primeng": "^8.0.3",
+    "rxjs": "^6.5.3",
     "tslib": "^1.10.0",
-    "typescript": "~2.5.3",
+    "typescript": "~3.5.3",
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.803.7",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular/language-service": "^8.2.9",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
     "codelyzer": "^5.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^4.3.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
+    "protractor": "^6.0.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1"
   },
