@@ -6,23 +6,23 @@ import { ProductService } from '../../../../services/product.service';
 import { Brand } from '../../../../domain/brand';
 
 @Component({
-  selector: 'app-add-brand',
-  templateUrl: './add-brand.component.html',
-  styleUrls: ['./add-brand.component.css']
+	selector: 'app-add-brand',
+	templateUrl: './add-brand.component.html',
+	styleUrls: ['./add-brand.component.css']
 })
 export class AddBrandComponent implements OnInit {
 
-	brand:Brand = new Brand();
+	brand: Brand = new Brand();
 
 	constructor(private brandService: BrandService,
-							private router: Router,
-							private productService: ProductService,
-							private toastr: ToastrService,
-							private route: ActivatedRoute) { }
+		private router: Router,
+		private productService: ProductService,
+		private toastr: ToastrService,
+		private route: ActivatedRoute) { }
 
-  ngOnInit() {
+	ngOnInit() {
 		this.loadBrandToEdit();
-  }
+	}
 
 
 	saveBrand() {
@@ -31,13 +31,13 @@ export class AddBrandComponent implements OnInit {
 			this.router.navigate(['/admin/brands']);
 		}, error => {
 			this.toastr.error(String(error));
-		})
+		});
 	}
 
 	loadBrandToEdit() {
-		let id = this.route.snapshot.paramMap.get('id');
+		const id = this.route.snapshot.paramMap.get('id');
 		if (id) {
-			this.brandService.getBrand(id).take(1).subscribe((result: Brand) => {
+			this.brandService.getBrand(id).subscribe((result: Brand) => {
 				this.brand = result;
 			}, error => {
 				this.toastr.error(String(error));
