@@ -1,11 +1,11 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '../../../../services/user.service';
+import { ActivatedRoute, Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
-import { RoleService } from '../../../../services/role.service';
-import { Router, ActivatedRoute } from '@angular/router';
-import { User } from '../../../../domain/user';
 import { Account } from '../../../../domain/account';
 import { Role } from '../../../../domain/role';
+import { User } from '../../../../domain/user';
+import { RoleService } from '../../../../services/role.service';
+import { UserService } from '../../../../services/user.service';
 
 @Component({
   selector: 'app-user-form',
@@ -15,15 +15,15 @@ import { Role } from '../../../../domain/role';
 export class UserFormComponent implements OnInit {
 
   roles: Role[] = [];
-  // user = { account :{ username:"", password: ""}};
+  // user = { account :{ username:'', password: ''}};
   user: User = new User();
-  confirmPWD: string = "";
+  confirmPWD = '';
 
   constructor(private userService: UserService,
-              private roleService: RoleService,
-              private router: Router,
-              private route: ActivatedRoute,
-              private toastr: ToastrService) { }
+    private roleService: RoleService,
+    private router: Router,
+    private route: ActivatedRoute,
+    private toastr: ToastrService) { }
 
   ngOnInit() {
     this.user.setAccount(new Account());
@@ -33,11 +33,11 @@ export class UserFormComponent implements OnInit {
 
   register() {
     this.userService.registerUser(this.user).subscribe((result: any) => {
-      this.toastr.success("User added successfully");
+      this.toastr.success('User added successfully');
       this.router.navigate(['/admin/users']);
     }, error => {
-      this.toastr.error("Error occured while creating the user");
-    })
+      this.toastr.error('Error occured while creating the user');
+    });
   }
 
   loadRoles() {
@@ -45,29 +45,29 @@ export class UserFormComponent implements OnInit {
       this.roles = result;
     }, error => {
       this.toastr.error(String(error));
-    })
+    });
   }
 
   loadUserToEdit() {
-    let id = this.route.snapshot.paramMap.get('id');
-    if(id) {
+    const id = this.route.snapshot.paramMap.get('id');
+    if (id) {
       this.userService.findById(id).subscribe((result: User) => {
         this.user = result;
         console.log(this.user.bDate);
         // set selected role. Again, there has to be a better way to do this!
-        if(this.roles) {
+        if (this.roles) {
           const p = this.user;
           let match;
-          this.roles.forEach(function(br){
-            if(br.id == p.role.id){
-                match = br;
+          this.roles.forEach(function (br) {
+            if (br.id === p.role.id) {
+              match = br;
             }
-          })
+          });
           this.user.role = match;
         }
       }, error => {
         this.toastr.error(String(error));
-      })
+      });
     }
   }
 
