@@ -1,27 +1,24 @@
 import { Injectable } from '@angular/core';
-import { CanActivate } from '@angular/router/src/interfaces';
-import { AuthenticationService } from './../services/authentication.service';
-import { Router, RouterStateSnapshot, ActivatedRouteSnapshot } from '@angular/router';
-import { JwtHelper } from 'angular2-jwt';
+import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from '@angular/router';
+import { JwtHelperService } from '@auth0/angular-jwt';
 
 @Injectable()
 export class AuthGuard implements CanActivate {
 
-  constructor(private authenticationService:AuthenticationService, private router:Router) { }
+  constructor(private router: Router) { }
 
   canActivate(router: ActivatedRouteSnapshot,
-              state: RouterStateSnapshot){
+    state: RouterStateSnapshot) {
     let activate = false;
-    let jwtHelper = new JwtHelper();
-    var token = localStorage.getItem('token');
-    if(token){
+    const jwtHelper = new JwtHelperService();
+    const token = localStorage.getItem('token');
+    if (token) {
       activate = !jwtHelper.isTokenExpired(token);
     }
-    
-    if(activate)
+    if (activate) {
       return activate;
-
-    this.router.navigate(['/login'], {queryParams: {src: state.url}});
+    }
+    this.router.navigate(['/login'], { queryParams: { src: state.url } });
     return false;
   }
 }
